pyproject.toml
[project]
name = "bridge-cli"
version = "0.1.0"
description = "A CLI tool to analyze technical debt in GitHub repositories"
authors = []
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "click",
    "typer",
    "rich",
    "requests",
    "radon",
    "bandit",
    "python-dotenv",
    "pyfiglet",
    "GitPython"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.optional-dependencies]
test = ["pytest>=7.0"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[project.scripts]
bridge = "bridge_cli.cli:cli" 